% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jcutil-package.R
\docType{package}
\name{jcutil-package}
\alias{jcutil}
\alias{jcutil-package}
\title{jcutil: Jamie package}
\description{
Utility functions for multi-collinearity.
}
\author{
\strong{Maintainer}: James Chamberlain \email{Jamie.Chamberlain@cancervic.org.au} (YOUR-ORCID-ID)

}
\keyword{internal}
