% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{make_shades}
\alias{make_shades}
\title{Make shades}
\usage{
make_shades(colour, n, lighter = TRUE)
}
\arguments{
\item{colour}{The colour to make shades of.}

\item{n}{The number of shades to make}

\item{lighter}{Whether to make it lighter (\code{TRUE}) or darker (\code{FALSE})}
}
\value{
A vector of \code{n} colour hex codes
}
\description{
Given a colour make n lighter or darker shades.
Very interesting package.
}
\details{
this is the details of the function
}
\examples{
# Five lighter shades
make_shades("goldenrod", 5)
# Five darker shades
# note: \dontrun or \dontshow could be used
make_shades("goldenrod", 5, lighter=FALSE)
}
