% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{plot_colours2}
\alias{plot_colours2}
\title{make colours and plot them, both lighter and darker}
\usage{
plot_colours2(colour, numshades)
}
\arguments{
\item{colour}{base colour for plot (string)}

\item{numshades}{number of shades}
}
\description{
make colours and plot them, both lighter and darker
}
